+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Yii Framework 2 redis extension Change Log
 - Enh #81: Allow setting `Connection::$database` to `null` to avoid sending a `SELECT` command after connection (cebe)
 - Bug #82: Fixed session object destruction failure when key expires (juffin-halli, samdark)
 - Bug #93: Fixed `yii\redis\ActiveRecord::deleteAll()` with condition (samdark)
+- Enh #53: Added Redis based Mutex `\yii\redis\Mutex` (turboezh)
 
 
 2.0.5 March 17, 2016